CustomFields refactor components to base folder
