+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "imask": "^7.1.3",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-datepicker": "^4.16.0",
