+++ src/components/@ui/CustomFields/index.js
@@ -1,0 +1,7 @@
+import CustomSwitch from "./CustomSwitch";
+import CustomTextField from "./CustomTextField";
+
+export {
+  CustomSwitch,
+  CustomTextField
+};
